Selection and checker method
The first selection method will be used to get the initial units and the second selection method will be used to get the final units.

The sameOptionsChecker method will be what checks if the user has used 2 of the same options from the selection. From there, they can prompted to make a different input.